style(rubocop): resolve Rails/LexicallyScopedActionFilter by inlining SponsorConcerns
The cop requires that `before_action` `only:` actions be explicitly defined on
the class where the filter is declared. `Admin::SponsorConcerns` was only
included by `Admin::WorkshopsController`, yet the `sponsor`, `destroy_sponsor`,
`host`, and `destroy_host` actions were defined inside a nested
`InstanceMethods` module — invisible to the cop.

Inlined the concern into the controller (YAGNI — no other controller used it).
Also replaced two `update_attribute` calls with `update` since the moved code
was no longer covered by the SkipsModelValidations exclusion.
